Change - don't backup decommissioned TeamCity agent files
When ensuring that a specific TeamCity Agent release is to be absent
(removed), don't bother making backups of any of the Agent files.  A new
installation can be ~700 MB and one that's been in service for a long
time may have several GB of logs, etc.  The puppet backup procedure
requires that each of the files be sent back to the Puppet Master over
HTTP, a rather slow process that has little to no benefit here.