Change - improve accuracy of parameter documentation
Because the "content" and "source" parameters are passed to a File-type
resource these behave in the exact same way.  The File resource does not
require either of these.  Passing both as undef is the same as passing
neither.  The File resource will only manage those aspects that are not
undef, so the template documentation here should better reflect that.