Bug - teamcenter-sync continues if target not mounted
A series of events led to some surprising discoveries:

    - puppet stopped running on tc-util after the CA certificate expired
      because the tc-util certificate never got signed by the new CA

    - because puppet wasn't running it was possible for remote file
      systems to become unmounted and stay that way

    - because the remote file systems were unmounted a TARGET would
      ultimately be just the root filesystem which meant:

        - the real TARGET did not actually get updated

        - the root file system free space would be effectively exhausted
          completely

The script has been revised to abort if either the SOURCE or TARGET
remote file systems are not mounted.  In the case of TARGET, this
prevents the situation described above.  In the unseen case of SOURCE,
this prevents something much worse -- synchronizing an empty directory
to an otherwise good TARGET.