Change - rewrite Class[koji::database]
This was partly required due to Dart's adoption of the
puppetlabs-postgresql module having replaced the original
doubledog-postgresql module we used.  The code here was not maintained
during that transition.

This class now leverages the capabilities of the replacement module in
that it fully prepares the PostgreSQL database for Koji use.
Specifically, it now not only installs the PostgreSQL server package and
configures authentication, but also:

    - creates required roles
    - imports the Koji database schema
    - bootstraps the Koji administrator account

Furthermore, this class has received a number of other improvements:
    - greatly improved documentation
    - puppet-lint conformance
    - extensive control via parameters whose defaults reflect the Fedora
      Project's standards