Change - refresh koji module
This module has been unmaintained since an evaluation of Koji was done
on Fedora 19.  At that time it didn't appear to fit our needs for
a continuous integration system.  TeamCity now occupies that role, but
is extremely weak at building RPMs, which is exactly what Koji excels
at.  Thus Koji is being resurrected at Dart to be integrated with
TeamCity.  This commit represents a variety of improvements to bring the
module up to usable standards for Fedora 21 and to follow current best
practices.  Specifically, in no particular order:

- all config files now have a "managed by Puppet" header
- all config files have been synchronized with their upstream origins
- improved documentation throughout; now using RDoc format
- all classes with service now provide ensure/enable parameters
- all references now use absolute namespaces to avoid nasty surprises
- package events trigger service refreshes
- puppet-lint compliance
- mash_repo resources can now have an arbitrary name independent of the
  name actually given to the repo file
- parameterized Koji administrator account name
- minor resource consolidation