Bug - koji::database doesn't assign permissions
A critical record was not inserted into the user_perms table for the
administrator account.  This would have required another exec resource
for the Puppet class along with yet another pair of log and flag files
to remain idempotent.  This commit introduces a new shell script
(created from a template) to handle those Koji database bootstrapping
tasks that cannot be handled directly by the puppetlabs-postgresql
module.