Change - improve prompts and defaults for certgen.sh
This critical script assists with X509 certificate creation for the Koji
services and users.  The Common Name field is critical for correct
operation of Koji's authorization scheme.  The Organization Unit field
isn't critical to the operation but is fundamental for understanding
a certificate's significance.  These improved prompts and usually
correct defaults should make using this tool substantially easier since
Koji's requirements aren't clearly documented and debugging faults can
be time consuming, if not challenging.  This is especially important
given the infrequent use of this tool.