New - host mdct-koji-f21 obsoletes mdct-koji (f19)
This commit drops all resources specific to the original mdct-koji host
that was built on Fedora 19 for evaluation purposes and replaces them
with equivalent or improved variants that are applicable to a build on
Fedora 21 and against the much improved Puppet module for Koji.

This does not yet include support for mash nor has anything been vetted
out here.  Validation has only gone so far as proving the services start
correctly although to get that far some components such as the Koji
Certificate Authority, Koji Hub, Koji CLI and Koji web interface have
already seen quite a bit of use.  To be sure, no builds have yet passed
through the system though.