New - hint at undocumented kojid configuration options
It's hard to say whether these are to be considered for "internal use"
or some such only or if they are fully intended for use but simply not
documented since the only documentation that exists for any of these
options are what appears in the config file by way of commentary.  Since
I had delved deep enough into the code for option parsing it only made
sense to add mention of these to the configuration template with a big
fat disclaimer.  Having awareness of these might prove rather valuable
in the future.