Change - improve documentation of koji::mash::repo::repoview_url
The original documentation for this parameter was derived from an
example mash configuration file.  However a closer review of the actual
mash code reveals that this parameter simply passes into the repoview
command-line as the argument to the -u option.  Thus the documentation
for this parameter now reflects the repoview -u documentation verbatim.