New - mash-everything script
Class[koji::mash] now deploys a "mash-everything" script which
iteratively mashes each configured repository and composites them into
a super-repository.  The repositories that mash-everything iterates over
are configured via the mash-everything.conf file which itself is
populated via each usage of Define[koji::mash::repo].  This means the
configuration is fully dynamic.

After mash-everything has composited the super-repository that is used
to synchronize the public copy of the same.  This synchronization step
is necessary to minimize disruptions to the public repository since the
mash utility completely destroys its target -- the process it employs is
one of delete-then-create rather than any kind of update-in-place.  The
net effect of mash-everything is to hide this behavior and limit the
disruption to the equivalent of using any other Fedora mirror.