New - manage Apache httpd on mdct-00fs
This commit accomplishes several things, starting with the real goal
here followed by things that had to be done along the way.

1.  Configure httpd to serve our master (by fiat) Git repositories via
Git's Smart HTTP.  This link will be used for, at minimum, Koji builds
from SCM which must be able to clone the repositories.  Doing so with
SSH keys would require careful administration that would remain rather
hidden and obscure, i.e., not ideal.

2.  httpd has never been managed by Puppet prior to this.  It was
introduced with much attention to the changes implemented.  Excepting
specific changes listed in this commit, the only other changes
introduced by Puppet here were to comments in the configuration files
and these all appeared to be limited to whitespace.

3.  mdct-00fs now serves our /pub directory structure through
http://mdct-00fs/pub.  This had been previously available only via
http://mdct-00fs/ftp/pub which is inconsistent with all of our other
hosts serving this content.  The latter should now be considered
deprecated.

4.  The "ftp" symlink (http://mdct-00fs/ftp) predates my involvement
with MDC.  I don't know exactly why it was set up this way, but I do
know for certain there are many extant references embedded in other
applications.