Bug - Koji-Web certificates are unmanaged
While the Koji-Web component was functional, it could be considered bad
practice for it to source its certificates directly from the Certificate
Authority.  This almost certainly got overlooked during the initial
struggle to get Koji operational -- a shortcut that never got cleaned up
until now.

This revision places a copy of the Koji-Web component's client
certificate within the "dart" module alongside all the other Koji
infrastructure certificates.  From here Puppet can push this to the host
along with the CA certificates that signed it and the Koji-Hub's client
certificate.  In effect, Puppet now fully manages the SSL certificates
for the Koji-Web component.