Refactor - formally declare Koji Web and Kojira hosts
Because the Sigul integration with Koji requires that the Sigul Bridge
share some common configuration with the Koji Web service, the dynamic
naming of the Puppet resources previously used was at risk for breakage
in the future should either the Web host or the Bridge host be given
a new hostname.  The Puppet resources are still dynamically named for
all the same reasons as before, but are now done in such a way as to
eliminate this risk by making the use of this sharing explicit and
traceable with simple 'git grep' operations.  While only
Class[dart::subsys::koji::web] really needed to be changed here, the
concept is just as apropos to Kojira so both are equally treated.

In case it's less than obvious, $fqdn in the "client_cert" would always
resolve correctly for the Koji-Web host, but the Sigul-Bridge host would
not have been able to rely on the same dynamic expression using $fqdn
because it is not intended to point at itself, but at the Koji-Web host.