New - integration of koji into sigul
This extends the management of the Sigul Bridge with Koji configuration
and SSL certificates for authenticating the client and server
components.  This is required for automated signing of RPMs.  At this
stage, I believe the Sigul Bridge is capable of retrieving unsigned RPMs
from the Koji Web interface, queuing them for signing by the Sigul
Server and pushing the signed RPMs back into Koji.  However, all testing
at this point has been through manual requests via the Koji CLI and
getting the signed RPMs out to the mashed repository has required
additional Koji commands.  Thus if my summary and understanding is
correct, all that remains is automating the processes with the tools in
the koji-hub-plugins package.  There may also be special work required
to get signed the packages already built from when I was testing and
smoothing out the build processes.