Bug - dart::subsys::koji classes not portable
Class[dart::subsys::koji::cli] is the ideal way to manage the Koji CLI
on any host so it should not use the $fqdn fact in setting the "web"
parameter.  Instead it should use the value of
dart::subsys::koji::params::web_host so this class is more portable.

While Class[dart::subsys::koji::hub] is unlikely to be used anywhere but
on the Koji Hub itself, it is also being changed here for much the same
reasons.  If nothing else, it is less surprising and makes the $web_host
value more authoritative.