Bug - mash-everything retains untagged builds
If a build has been untagged mash-everything would not have that rpm
appearing in the associated yum repository if the repository were being
created from scratch.  When mash-everything was updating an existing yum
repository however, the rpm would be retained simply by being ignored.
The script now deletes objects in the yum repository that would not be
present if the yum repository was being created anew.