Bug - Sigul key maps need unique names
The original data structure was flawed.  While it would be nice to use
key name as the index to the hash, that precludes reuse of the signing
key for more than Koji target.  Since all of our packages to date,
regardless of the Fedora release to which they're associated, have been
signed with the same key, it's desirable to reuse the same key for at
least our Fedora 20 and Fedora 21 targets -- to remain backwards
compatible.  New keys will likely be issued for upcoming releases to
follow the Fedora Project's practice.

This revision moves the key name into a new hash element and moves the
inner "tag" key the index value for each map entry.  The structure of
key_map could be better normalized to avoid the silly looking duplicates
in the dynamically constructed dictionary within sigulsign_unsigned.py,
but to do so would make the Puppet class much less user-friendly.