New - "interval" parameter for Class[sigul::auto_signer]
The frequency for which the build signing process occurs is now
controllable via this option.  The default is the fasted supported at
once per minute, but it might be useful to reduce this for e.g., when
faced with setup challenges to avoid excessive mailings from cron for
failed jobs.