Refactor - /pub is now reachable without /ftp prefix
The web server on mdct-00fs has been reconfigured to serve the /pub
filesystem such that it is no longer necessary to prefix this with /ftp
as long had been the case.  Now that mdct-00fs is now consistent with
all other hosts, this commit revises a number of URL references to
remove the unnecessary /ftp prefix.