Change - don't clobber repoview with mash-everything
Within individual yum repositories, mash will completely rebuild the
repoview directories -- that's all fine and good.  However, there may be
a repoview directory within the REPO_DIR (i.e., the root of all the yum
repositories) which may contain an index page to the others.  This
repoview directory should be left unharmed and as is.