Change - update config for puppet-tools-1.2
puppet-tools is now on the latest release from doubledog.org and the
configuration requirements are significantly different due to the
addition of several new tools (e.g., puppet-module-report,
puppet-dlo-cleanup) as well as improvements to the existing tools.  The
changes made here are necessary for puppet-tools to remain functional in
any capacity after this upgrade.