New - automatically regenerate Koji repo metadata
This establishes a cron job, a "regen-repos" script along with
a configuration file, all of which exist for the sole purpose of
detecting when Koji's external repositories mutate (e.g., Fedora
updates) so that Koji's internal repository metadata can be regenerated
thereby avoiding failed builds that would be inevitable otherwise.

The cron job ensures this is run regularly, every 10 minutes.  The
configuration file provides a clean interface between the Puppet
declarative and the scripts need for direction.  The configuration file
accomplishes this by associating the Koji tags representing various
buildroot environments with file system paths that provide the Koji
external repositories content.  The script can the simply iterate over
each tag and determine if the external content has changed since the
prior execution.