Bug - access forbidden to ISO images created by Koji
This resolves errors such as the following when a user attempts to
download an ISO image that has been created by "koji spin-livecd":

    AH00037: Symbolic link not allowed or link target not accessible:
    /mnt/koji/work/tasks/1140/1140/jfspin1-6.0-1.iso

To resolve this, httpd must be permitted to traverse the symlink.  For
the example above, this looks like:

    /mnt/koji/work/tasks/1140/1140/jfspin1-6.0-1.iso ->
    /mnt/koji/scratch/d13677/task_1140/jfspin1-6.0-1.iso

This revision now permits httpd to follow such symlinks.  To keep this
as secure as possible, SymLinksIfOwnerMatch has been used instead of
FollowSymLinks.  That has also required that the "apache" user/group
take ownership of top_dir itself too, not just the content below.

See also the relevant discussion per:

    http://comments.gmane.org/gmane.linux.redhat.fedora.buildsystem/4222