Change - dart-ntp ==> ringingliberty-chrony
This commit migrates us from the dart-ntp module to the much better
ringingliberty-chrony module and at the same time moves conditional NTP
parameters out of Class[dart::abstract::base_node] and into Hiera for
a polymorphic solution.

The resulting configuration for chronyd does not differ much as result,
but here are the highlights of what has been observed:

    - loss of in-line comments; some were of dubious value
    - NTP serving is now turned off for hosts where it should be
      client-only; this follows a Fedora policy implemented several
      releases ago which we failed to follow until now
    - makestep is now configured as Fedora default; we had customized
      this to use "makestep 100 3", but for undocumented reasons

Another notable consequence of this migration is that we no longer
support the legacy ntpd.  The dart-ntp module would support both ntpd
and chronyd, choosing one based on the Fedora release.  This should not
be much of an issue because there should be few, if any, hosts still
running Fedora 15 or earlier AND which are managed by Puppet.