Change - move inclusion of puppet::client into base_node
This had become an unmanageable mess.  Long ago the inclusion of
puppet::client was moved out of the abstract::base_node and into those
classes which inherited it so as to prevent the client for a few select
hosts.  I believe it was only PICAPS servers (test and production) that
desired to prevent the client from continued running and state
enforcement.  However, the picture has become nearly impossible to
comprehend now.

This revision brings the puppet::client class inclusion back into
dart::abstract::base_node so that it's universally applied everywhere.
For any host not desiring that, they should seek an explicit, per-host
override via Hiera.