Change - include sendmail on mdct_00fs
This is part of reconciliation effort with hopes of having our file
server node class actually inherit from dart::abstract::base_node, at
least.

This has no functional effect.  This does include a patch to the
sendmail module to eliminate sm-client from the list of services when
applied to Fedora 14 or older.  Previously, sm-client was in the list of
services unconditionally, but F14 reports:

    sudo service sm-client start
    sm-client: unrecognized service