Bug - NFS ID mapper cannot be enabled on Fedora 21
As of Fedora 21, the NFS ID mapper service has become a static unit (in
systemd parlance).  It is automatically "pulled in" and started by
nfs-server.service.  For clients however, it is now considered obsolete
having been replaced by nfsidmap.  There is no systemd unit for nfsidmap
that needs to be managed.  There is only the requisite configuration to
request-key and Fedora provides this in
/etc/request-key.d/id_resolver.conf for us already.

While the nfs-idmapd.service (for rpc.idmapd) is now static and can no
longer be enabled/disabled, it is possible to start/stop this service
unit.  Previously it was ensured to be "running" (and still is for older
Fedoras) so it's helpful to stop it to prevent any administrator
confusion in believing it still relevant.