Change - force MAC address on bridge interface
This is being done to prevent the MAC address on br0 from changing every
time 'brctrl addif' is executed by run-iso on our (Daniel and John)
workstations.

It seems that brctrl has a policy of setting the bridge's MAC to the
smallest MAC value of all member interfaces, unless a MAC is
administratively assigned to the bridge interface.  By administratively
assigning one here, this will prevent the run-time change which is
undesirable mostly because it causes puppet to rewrite the ifcfg-br0
file for the setup and again for the tear-down, which consequently makes
for a fair bit of audit report noise.

The MAC addresses chosen here rests in one of the Locally Administered
Address Ranges[0] and embeds our Dart employee IDs in a scheme to make
them identifiable and unique-ish.

[0] https://serverfault.com/questions/40712/what-range-of-mac-addresses-can-i-safely-use-for-my-virtual-machines/