Bug - redundant exec of repoview
The dart-repoview script does not need to run repoview since that is
being handled by mash.  It should continue to manage the index from
which the user selects a particular repoview by OS release and
architecture.  That's the real job of this script anymore.

(Furthermore, this was executing repoview for the old "mdct"
repositories rather than the new "dart" ones.)