Bug - packages promoted too quickly left unsigned
Unsigned packages will be sought out within dart-testing repositories
now in addition to dart-candidates.  It's not uncommon for freshly built
packages to be rapidly promoted to testing.  Theoretically dart-released
might also be listed here, but dubious are the claims of actual testing
of any package promoted to released from testing (or, god forbid,
candidates) in less than 60s (the auto-signing interval).