Change - impose shorter herd ssh conn timeout
By default the OpenSSH client relies only on the OS's TCP connection
timeouts.  This is excessively long for use cases of "herd pssh" where
the node is rebooted and reacquiring the connection in quick order is
expected.  This revision causes the client to impose its own timeout of
10s so that herd can reattempt the connection.  This had a few distinct
advantages:

    - Relying only on the TCP timeout might result in packets being sent
      out to establish (or renew) a connection just prior to the node's
      interface coming online.  Waiting for the TCP timeout is likely to
      take longer than just giving up and reattempting when this next
      attempt can succeed.

    - This gives herd an earlier chance to recognize a node changing IP
      addresses.

I'm applying this to myself only for now to evaluate the overall UX.
The 1st point has already been proven, the 2nd is almost guaranteed so
the real risk here is unanticipated side effects.  If none are observed,
this should become a global default.