New - enable the Koji Garbage Collector
This revision causes our Koji server to also act as a garbage collector
to eliminate older builds that no longer need to be retained due to
a policy, which at present:

    - Preserves builds tagged for the any of the *-testing repositories
      so that these builds are retained for up to 4 weeks, which should
      provide developers adequate time to evaluate their build, even if
      it has been superseded by newer builds.

    - Moves into the "trash can" builds that have been obsoleted by, at
      minimum, 3 newer builds (excepting those described above).

    - Permanently destroys builds that have been in the "trash can" for
      at least 2 weeks.

This revision also establishes a new X509 certificate for a new
"koji-gc" user.  This user was manually added to the Koji DB and granted
the "admin" permission.