New - serve old Fedora releases via NFS
In order to make select releases of Fedora <= 20 available via /pub as
before, the old_fedoras filesystem is now being exported via NFS on
mdct-00ut.  This will provide us with continuity for Fedora 18 and 20
for the immediate future as these are still in use by AOS 5.2, 6.0 and
PICAPS servers.