Change - mock to use "dart" repos for Fedora >= 20
We transitioned from the legacy "mdct" repositories to their
contemporary equivalent "dart" repositories circa Fedora 20/21.  With
those two releases there was actually overlap but the newest builds for
either (and later releases, of course) occurred only in the "dart"
repositories.

This revision adapts mock to use the newer repositories when the build
target is Fedora 20 or newer.  This also merges in some upstream changes
to the configuration (based on a Fedora 23 example) that:

    - enables GPG package signature checking (for Fedora packages).
      (There's no good way to do so for Dart packages since the
      signatures are kept in /etc/mock/pki which is owned by the mock
      package -- and needs a pki.d for others).
    - revises some caching and debug level options
    - adds settings for dnf (e.g., week deps)