New - Add puppetboard module to manage puppetboard installs.
Puppetboard is a web interface to PuppetDB aiming to replace
the reporting functionality of Puppet Dashboard.

Puppetboard relies on the pypuppetdb library to fetch data from
PuppetDB and is built with the help of the Flask microframework.

This module automatically installs the necessary python modules
using pip and also configures the apache virtual host needed
to run the application.